refactor: refactor index to smaller files
